[TASK] Refactor JobConfigurationRepository
